+++ lib/application_enumeration.rb
@@ -100,4 +100,33 @@ module ApplicationEnumeration
     a
   end
 
+  def self.relation_targets_community?(relation)
+    case relationship_type(relation)
+    when :has_many
+      relation.class_name.safe_constantize.is_community?
+    when :has_one
+      raise TaxonWorks::Error, "Has one support not implemented in unify, throw eggs at the devs."
+    when :belongs_to
+      if k = relation.options[:class_name]
+        k.safe_constantize.is_community?
+      else
+        raise TaxonWorks::Error, "Missing attribute class_name on #{relation.name}."
+      end
+    end
+  end
+
+  # collection?, has_mone?  belongs_to?
+  def self.relationship_type(relation)
+    if relation.collection? # class.name.match('HasMany')
+      return :has_many
+    elsif relation.has_one? # class.name.match('HasOne')
+      return :has_one
+    elsif relation.belongs_to? # class.name.match('BelongsTo')
+      return :belongs_to
+    end
+
+    raise TaxonWorks::Error, "Unknown relationship type for #{relation.name}."
+  end
+
+
 end