+++ CHANGELOG.md
@@ -7,7 +7,31 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ## [unreleased]
 
-\-
+### Added
+
+- Created/updated overviews for user data
+- Unify objects task [#970]
+- Attribution to ObservationMatrix
+- New biological association task [#4026]
+- Duplicate OTU predictor task [#4083]
+
+### Changed
+
+- Local identifier display in CollectionObject tag now prefers position to break tie with RecordNumber and CatalogNumber [#4074]
+
+### Fixed
+
+- Simple TaxonName batch load failing with invalid children
+- `api/v1/images/:id` broken for non-integer ids
+- Containerizing objects prevented identific increments
+- New dichotomous key: Radial annotator is not loading the correct data [#4076]
+
+[#4083]: https://github.com/SpeciesFileGroup/taxonworks/issues/4083
+[#4026]: https://github.com/SpeciesFileGroup/taxonworks/issues/4026
+[#4074]: https://github.com/SpeciesFileGroup/taxonworks/issues/4074
+[#4076]: https://github.com/SpeciesFileGroup/taxonworks/issues/4076
+[#4077]: https://github.com/SpeciesFileGroup/taxonworks/issues/4077
+[#970]: https://github.com/SpeciesFileGroup/taxonworks/issues/970
 
 ## [0.44.3] - 2024-10-03